+++ src/api/java/com/codetaylor/mc/artisanworktables/api/internal/config/IModuleWorktablesConfig.java
@@ -3,6 +3,8 @@ package com.codetaylor.mc.artisanworktables.api.internal.config;
 import com.codetaylor.mc.artisanworktables.api.internal.reference.EnumTier;
 import com.codetaylor.mc.artisanworktables.api.internal.reference.EnumType;
 
+import javax.annotation.Nullable;
+
 public interface IModuleWorktablesConfig {
 
   boolean enableWorktables();
@@ -20,6 +22,20 @@ public interface IModuleWorktablesConfig {
   int getFluidCapacityWorkshop(String tableName);
 
   int getTextHighlightColor(String tableName);
+
+  int getFluidTankOverlayColor(String tableName);
+
+  @Nullable
+  Integer getCraftingGridSlotBackgroundColor(String tableName);
+
+  @Nullable
+  Integer getFluidTankBackgroundColor(String tableName);
+
+  @Nullable
+  Integer getMainOutputSlotBackgroundColor(String tableName);
+
+  @Nullable
+  Integer getPlayerInventorySlotBackgroundColor(String tableName);
 
   boolean isTierEnabled(EnumTier tier);